big changes structurally to separate some logic
Separates the logic with some components moves most of the configuration like data to separate config JSON files.
