+++ src/configs/SandboxLocationRegionalItems.js
@@ -1,0 +1,13 @@
+const SandboxLocationRegionalItems = () => ([
+  'Northeast',
+  'Southeast',
+  'Midwest',
+  'Northern Great Plains',
+  'Northwest',
+  'Southwest',
+  'Southern Great Plains',
+  'Alaska',
+  'Hawaii',
+  'Puerto Rico'
+]);
+export default SandboxLocationRegionalItems;
