+++ src/configs/SandboxPeriods.js
@@ -1,0 +1,5 @@
+const SandboxPeriods = () => ([
+  '1900-current',
+  '1950-current'
+]);
+export default SandboxPeriods;
