+++ src/configs/SandboxRegionItems.js
@@ -1,0 +1,6 @@
+const SandboxRegionItems = () => ([
+  'National',
+  'Regional',
+  'State'
+]);
+export default SandboxRegionItems;
