+++ src/configs/SandoxLocationNamesHumanReadable.js
@@ -1,0 +1,255 @@
+const SandoxLocationNamesHumanReadable = () => ([
+  {
+    value: 'AK',
+    pullDownText: 'Alaska'
+  },
+  {
+    value: 'AL',
+    pullDownText: 'Alabama'
+  },
+  {
+    value: 'AZ',
+    pullDownText: 'Arizona'
+  },
+  {
+    value: 'AR',
+    pullDownText: 'Arkansas'
+  },
+  {
+    value: 'CA',
+    pullDownText: 'California'
+  },
+  {
+    value: 'CO',
+    pullDownText: 'Colorado'
+  },
+  {
+    value: 'CT',
+    pullDownText: 'Connecticut'
+  },
+  {
+    value: 'DE',
+    pullDownText: 'Delaware'
+  },
+  {
+    value: 'FL',
+    pullDownText: 'Florida'
+  },
+  {
+    value: 'GA',
+    pullDownText: 'Georgia'
+  },
+  {
+    value: 'HI',
+    pullDownText: 'Hawai\'i'
+  },
+  {
+    value: 'ID',
+    pullDownText: 'Idaho'
+  },
+  {
+    value: 'IL',
+    pullDownText: 'Illinois'
+  },
+  {
+    value: 'IN',
+    pullDownText: 'Indiana'
+  },
+  {
+    value: 'IA',
+    pullDownText: 'Iowa'
+  },
+  {
+    value: 'KS',
+    pullDownText: 'Kansas'
+  },
+  {
+    value: 'KY',
+    pullDownText: 'Kentucky'
+  },
+  {
+    value: 'LA',
+    pullDownText: 'Louisiana'
+  },
+  {
+    value: 'ME',
+    pullDownText: 'Maine'
+  },
+  {
+    value: 'MD',
+    pullDownText: 'Maryland'
+  },
+  {
+    value: 'MA',
+    pullDownText: 'Massachusetts'
+  },
+  {
+    value: 'MI',
+    pullDownText: 'Michigan'
+  },
+  {
+    value: 'MN',
+    pullDownText: 'Minnesota'
+  },
+  {
+    value: 'MS',
+    pullDownText: 'Mississippi'
+  },
+  {
+    value: 'MO',
+    pullDownText: 'Missouri'
+  },
+  {
+    value: 'MT',
+    pullDownText: 'Montana'
+  },
+  {
+    value: 'NE',
+    pullDownText: 'Nebraska'
+  },
+  {
+    value: 'NV',
+    pullDownText: 'Nevada'
+  },
+  {
+    value: 'NH',
+    pullDownText: 'New Hampshire'
+  },
+  {
+    value: 'NJ',
+    pullDownText: 'New Jersey'
+  },
+  {
+    value: 'NM',
+    pullDownText: 'New Mexico'
+  },
+  {
+    value: 'NY',
+    pullDownText: 'New York'
+  },
+  {
+    value: 'NC',
+    pullDownText: 'North Carolina'
+  },
+  {
+    value: 'ND',
+    pullDownText: 'North Dakota'
+  },
+  {
+    value: 'OH',
+    pullDownText: 'Ohio'
+  },
+  {
+    value: 'OK',
+    pullDownText: 'Oklahoma'
+  },
+  {
+    value: 'OR',
+    pullDownText: 'Oregon'
+  },
+  {
+    value: 'PA',
+    pullDownText: 'Pennsylvania'
+  },
+  {
+    value: 'PR',
+    pullDownText: 'Puerto Rico'
+  },
+  {
+    value: 'RI',
+    pullDownText: 'Rhode Island'
+  },
+  {
+    value: 'SC',
+    pullDownText: 'South Carolina'
+  },
+  {
+    value: 'SD',
+    pullDownText: 'South Dakota'
+  },
+  {
+    value: 'TN',
+    pullDownText: 'Tennessee'
+  },
+  {
+    value: 'TX',
+    pullDownText: 'Texas'
+  },
+  {
+    value: 'UT',
+    pullDownText: 'Utah'
+  },
+  {
+    value: 'VT',
+    pullDownText: 'Vermont'
+  },
+  {
+    value: 'VA',
+    pullDownText: 'Virginia'
+  },
+  {
+    value: 'VI',
+    pullDownText: 'Virgin Islands'
+  },
+  {
+    value: 'WA',
+    pullDownText: 'Washington'
+  },
+  {
+    value: 'WV',
+    pullDownText: 'West Virginia'
+  },
+  {
+    value: 'WI',
+    pullDownText: 'Wisconsin'
+  },
+  {
+    value: 'WY',
+    pullDownText: 'Wyoming'
+  },
+  {
+    value: 'National',
+    pullDownText: 'National'
+  },
+  {
+    value: 'Northeast',
+    pullDownText: 'Northeast'
+  },
+  {
+    value: 'Southeast',
+    pullDownText: 'Southeast'
+  },
+  {
+    value: 'Midwest',
+    pullDownText: 'Midwest'
+  },
+  {
+    value: 'Northern Great Plains',
+    pullDownText: 'Northern Great Plains'
+  },
+  {
+    value: 'Northwest',
+    pullDownText: 'Northwest'
+  },
+  {
+    value: 'Southwest',
+    pullDownText: 'Southwest'
+  },
+  {
+    value: 'Southern Great Plains',
+    pullDownText: 'Southern Great Plains'
+  },
+  {
+    value: 'Alaska',
+    pullDownText: 'Alaska'
+  },
+  {
+    value: 'Hawaii',
+    pullDownText: 'Hawai\'i'
+  },
+  {
+    value: 'Puerto Rico',
+    pullDownText: 'Puerto Rico'
+  }
+]);
+export default SandoxLocationNamesHumanReadable;
